+++ src/main/java/com/allclear/tastytrack/domain/restaurant/entity/Restaurant.java
@@ -1,5 +1,6 @@
 package com.allclear.tastytrack.domain.restaurant.entity;
 
+import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import com.fasterxml.jackson.datatype.jsr310.deser.LocalDateTimeDeserializer;
@@ -47,6 +48,7 @@ public class Restaurant {
     private String lat;                  // 위도
 
     @Column(nullable = false)
+    @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
     @JsonSerialize(using = LocalDateTimeSerializer.class)
     @JsonDeserialize(using = LocalDateTimeDeserializer.class)
     private LocalDateTime lastUpdatedAt; // 최종 수정일자
